requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
jinja2==3.1.4
sqlalchemy==2.0.35
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
itsdangerous==2.2.0
psycopg[binary]==3.2.13
